+++ radiant/brush/FaceInstance.cpp
@@ -217,7 +217,7 @@ void FaceInstance::testSelect_centroid(Selector& selector, SelectionTest& test) 
 void FaceInstance::selectPlane(Selector& selector, const Line& line, PlanesIterator first, PlanesIterator last, const PlaneCallback& selectedPlaneCallback) {
 	for (Winding::const_iterator i = getFace().getWinding().begin(); i != getFace().getWinding().end(); ++i) {
 		Vector3 v(line_closest_point(line, (*i).vertex) - (*i).vertex);
-		double dot = getFace().plane3().normal().dot(v);
+		float dot = getFace().plane3().normal().dot(v);
 		if (dot <= 0) {
 				return;
 			}
